+++ src/object.js
@@ -187,8 +187,9 @@ export let omitBlank = x => _.omitBy(isBlank, x)
 export let omitEmpty = x => _.omitBy(_.isEmpty, x)
 
 // (f, g) -> (x, y) -> {...f(x, y), ...g(x, y)}
-export let mergeOverAll = fns =>
+export let mergeOverAll = _.curryN(2, (fns, ...x) =>
   _.flow(
     _.over(fns),
     mergeAllArrays
-  )
+  )(...x)
+)